# Performance testing using k6 for event handler