Related to #34
A load test was done on the cluster using [K6](https://grafana.com/docs/k6/latest/). A variable number of `VUs` are used to create a burst of events in the cluster. 

The number of `APP.ADD` events captured by YHS is counted.

### Tests 1: 
Time: **15s, 5 virtual users (VUs)** for concurrent submission of pods in the cluster
Pod Submitted: 450
Event processed by YHS: 550

### Tests 2: 
Time: **15s, 10 virtual users (VUs)** for concurrent submission of pods in the cluster
Pod Submitted: 690
Event processed by YHS: 790

### Tests 3: 
Time: **15s, 10 virtual users (VUs)** for concurrent submission of pods in the cluster
Pod Submitted: 909
Event processed by YHS: 909

### Tests 4: 
Time: **25s, 50 virtual users (VUs)** for concurrent submission of pods in the cluster
Pod Submitted: 4008
Event processed by YHS: 4108

<img width="1431" alt="image" src="https://github.com/G-Research/yunikorn-history-server/assets/32765701/91a3ddc0-e5d9-420f-8d22-804c4cc02464">